Fix firebase import & slight refactor
