+++ src/app/components/auth/login/login.component.js
@@ -1,28 +1,10 @@
 import loginTemplate from './login.html';
-import loginCtrl from './login.controller';
 
-const login = {
+const loginComponent = {
   templateUrl: loginTemplate,
-  controller: loginCtrl,
+  controller: function loginCtrl() {
+
+  },
 };
 
-angular
-  .module('components.auth')
-  .component('login', login)
-  .config(config);
-
-function config($stateProvider, $urlRouterProvider) {
-  $stateProvider
-    .state('auth.login', {
-      url: '/login',
-      component: 'login',
-    });
-
-  $urlRouterProvider.otherwise('/auth/login');
-}
-config.$inject = [
-  '$stateProvider',
-  '$urlRouterProvider',
-];
-
-export default login;
+export default loginComponent;
