--- src/app/components/auth/login/login.controller.js
@@ -1,5 +1,0 @@
-function loginCtrl() {
-
-}
-
-export default loginCtrl;
