+++ src/app/components/auth/login/login.html
@@ -3,7 +3,7 @@
 </div>
 
 <div class="auth__info">
-  <a ui-ref='auth.register'>
+  <a href="" ui-sref="auth.register">
     Don't have an account? Create one here.
   </a>
 </div>
