+++ src/app/components/auth/login/login.module.js
@@ -1,0 +1,23 @@
+import uiRouter from 'angular-ui-router';
+import loginComponent from './login.component';
+
+const login = angular
+  .module('components.auth.login', [
+    uiRouter,
+  ])
+  .component('login', loginComponent)
+  .config(config)
+  .name;
+
+function config($stateProvider) {
+  $stateProvider
+    .state('auth.login', {
+      url: '/login',
+      component: 'login',
+    });
+}
+config.$inject = [
+  '$stateProvider',
+];
+
+export default login;
