+++ src/app/components/auth/register/register.component.js
@@ -1,25 +1,10 @@
 import registerTemplate from './register.html';
-import registerCtrl from './register.controller';
 
-const register = {
+const registerComponent = {
   templateUrl: registerTemplate,
-  controller: registerCtrl,
+  controller: function registerCtrl() {
+
+  },
 };
 
-angular
-  .module('components.auth')
-  .component('register', register)
-  .config(config);
-
-function config($stateProvider) {
-  $stateProvider
-    .state('auth.register', {
-      url: '/register',
-      component: 'register',
-    });
-}
-config.$inject = [
-  '$stateProvider',
-];
-
-export default register;
+export default registerComponent;
