--- src/app/components/auth/register/register.controller.js
@@ -1,5 +1,0 @@
-function registerCtrl() {
-
-}
-
-export default registerCtrl;
