+++ src/app/components/auth/register/register.html
@@ -3,7 +3,7 @@
 </div>
 
 <div class="auth__info">
-  <a ui-ref='auth.login'>
-    Don't have an account? Create one here.
+  <a href="" ui-sref="auth.login">
+    Already have an account? Log in here.
   </a>
 </div>
